Refactor overview stats for better data handling
